fixed some warnings
git-svn-id: file:///home/svn/humanoids/lib/somatic/trunk@3836 4f504ea5-f356-0410-adec-f1b3679f3bb7

+++ include/somatic.h
@@ -73,7 +73,7 @@ static inline int somatic_msg_send_buf_stack( ach_channel_t *chan,
                                  _somatic_private_n,                    \
                                  _somatic_private_msg );                \
         assert( _somatic_private_n == _somatic_private_r );             \
-        ach_put( chan, _somatic_private_buf, _somatic_private_r );      \
+        ach_put( chan, _somatic_private_buf, (size_t)_somatic_private_r ); \
     })
